rag_module/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag_module"
version = "0.1.0"
dependencies = [
    "langchain>=0.1.0",
    "llama-index>=0.9.8",
    "openai>=1.1.0",
    "chromadb>=0.4.18",
    "python-dotenv==1.0.0",
    "tiktoken>=0.5.2",
    "numpy>=1.24.3",
    "pandas>=2.0.3"
] 